+++ test/testReplacer.ts
@@ -1,15 +1,14 @@
 import expect from 'expect';
-const rand = require('../src/rand.ts');
-const WeightedChoice = require('../src/weightedChoice.ts').WeightedChoice;
-const Replacer = require('../src/replacer.ts').Replacer;
-const noOp = require('../src/noOp.ts');
+import * as rand from '../src/rand';
+import { WeightedChoice } from '../src/weightedChoice';
+import { Replacer } from '../src/replacer';
 
 describe('Replacer', function() {
 
   beforeEach(function() {
     rand.setRandomSeed(0); // pin seed for reproducibility
   });
-  
+
   it('on call returns well-formed object', function() {
     let weights = [
       new WeightedChoice('foo', 40),
@@ -20,7 +19,7 @@ describe('Replacer', function() {
     expect(result.replacement).toBe('foo');
     expect(result.choiceIndex).toBe(0);
   });
-  
+
   it('has a useful toString', function() {
     let weights = [
       new WeightedChoice('foo', 40),
@@ -29,6 +28,6 @@ describe('Replacer', function() {
     let replacer = new Replacer(weights, 'identifier', true);
     expect(replacer.toString()).toBe(
       'Replacer{weights: WeightedChoice{choice: foo, weight: 40},'
-        + 'WeightedChoice{choice: bar, weight: 60}, identifier: identifier, isSilent: true}');
+      + 'WeightedChoice{choice: bar, weight: 60}, identifier: identifier, isSilent: true}');
   });
 });