+++ test/testStringUtils.ts
@@ -1,23 +1,22 @@
 import expect from 'expect';
 
-const _stringUtils = require('../src/stringUtils.ts');
-const lineAndColumnOf = _stringUtils.lineAndColumnOf;
+import { lineAndColumnOf } from '../src/stringUtils';
 
 
 describe('lineAndColumnOf', function() {
   it('should work at the first character', function() {
-    expect(lineAndColumnOf('a', 0)).toEqual({line: 1, column: 0});
+    expect(lineAndColumnOf('a', 0)).toEqual({ line: 1, column: 0 });
   });
 
   it('should count newline characters as the ending of a line', function() {
-    expect(lineAndColumnOf('a\nb', 1)).toEqual({line: 1, column: 1});
+    expect(lineAndColumnOf('a\nb', 1)).toEqual({ line: 1, column: 1 });
   });
-  
+
   it('should work at the start of a new line', function() {
-    expect(lineAndColumnOf('a\nb', 2)).toEqual({line: 2, column: 0});
+    expect(lineAndColumnOf('a\nb', 2)).toEqual({ line: 2, column: 0 });
   });
 
   it('should work at the end of the string', function() {
-    expect(lineAndColumnOf('a\nba', 3)).toEqual({line: 2, column: 1});
+    expect(lineAndColumnOf('a\nba', 3)).toEqual({ line: 2, column: 1 });
   });
 });